Added data for Zygarde + forms + appearance
Added data for Zygarde as a species and its forms, then programmed its appearance mechanics to be unique and different from all other Pokemon.  Cells appear first during extreme type appeal differences, then cores are attracted to them, the cores merge with the cells to become 10% and 50% and 100% and when complete form is reached all type appeal values are inverted.  Crazy stuff.